changes and cleanups requested
